backend/requirements.txt
fastapi
uvicorn
openai
pymupdf
apify-client
python-dotenv
python-multipart
